+++ index.html
@@ -229,5 +229,7 @@
     </div>
   </template>
 
+  <script src="js/main.js" type="module"></script>
+
 </body>
 </html>
